requirements.txt
fastapi
uvicorn[standard]
requests
pillow
numpy
